+++ src/autoload.php
@@ -6,6 +6,9 @@
  * @license     MIT, http://flightphp.com/license
  */
 
-require_once __DIR__.'/core/Loader.php';
+require_once __DIR__.'/Flight//Core/Loader.php';
 
-\flight\core\Loader::autoload(true, dirname(__DIR__));
+
+
+
+\Flight\Core\Loader::autoload(true, __DIR__);