+++ docs/CHANGELOG.md
@@ -1,5 +1,9 @@
 ## MMCP Changes from initial mikecao fork.
 
-* Added .gitignore file
-* Added docs folder for further documentation
+
+* Updated to PSR-0 & composer friendly. Framework now in src folder. 
 * Moved example index.php to example folder with example .htaccess
+* Added docs folder for further documentation
+* Added .gitignore file
+
+